requirements.txt
Flask
Flask_Cors
openai
requests
python-dotenv
